+++ src/components/editors/EditableColumnEditor/EditableColumnEditorsRegistry.tsx
@@ -1,15 +1,26 @@
 import { dateTime, dateTimeFormat } from '@grafana/data';
-import { DateTimePicker, InlineField, InlineFieldRow, InlineSwitch, Input, Select, TextArea } from '@grafana/ui';
+import {
+  DateTimePicker,
+  FileDropzone,
+  InlineField,
+  InlineFieldRow,
+  InlineSwitch,
+  Input,
+  Select,
+  TextArea,
+} from '@grafana/ui';
 import { NumberInput } from '@volkovlabs/components';
 import React, { ChangeEvent } from 'react';
 
 import { TEST_IDS } from '@/constants';
 import { ColumnEditorType } from '@/types';
 import {
+  applyAcceptedFiles,
   cleanPayloadObject,
   createEditableColumnEditorRegistryItem,
   createEditableColumnEditorsRegistry,
   formatNumberValue,
+  toBase64,
 } from '@/utils';
 
 import { DateEditor, QueryOptionsEditor } from './components';
@@ -264,5 +275,46 @@ export const editableColumnEditorsRegistry = createEditableColumnEditorsRegistry
         })),
       };
     },
+  }),
+  createEditableColumnEditorRegistryItem({
+    id: ColumnEditorType.FILE,
+    editor: ({ value, onChange }) => (
+      <InlineFieldRow>
+        <InlineField label="Accept" tooltip="Specify comma-separated file extensions or keep blank to allow any file">
+          <Input
+            value={value?.accept}
+            onChange={(event) => {
+              const accept = event.currentTarget.value;
+              onChange(cleanPayloadObject({ ...value, accept }));
+            }}
+            placeholder="e.g., image/jpeg, application/pdf"
+            {...TEST_IDS.editableColumnEditor.fieldFileAcceptTypes.apply()}
+          />
+        </InlineField>
+      </InlineFieldRow>
+    ),
+    control: ({ onChange, config, isSaving }) => (
+      <InlineField label="File Upload" disabled={isSaving} grow={true}>
+        <FileDropzone
+          options={{
+            accept: applyAcceptedFiles(config.accept),
+            multiple: false,
+            onDrop: (files: File[]) => {
+              /**
+               * base64 result return
+               */
+              toBase64(files[0]).then((result) => {
+                onChange(result);
+              });
+            },
+          }}
+          {...TEST_IDS.editableCell.fieldFile.apply()}
+        />
+      </InlineField>
+    ),
+    getControlOptions: ({ config }) => ({
+      type: ColumnEditorType.FILE,
+      accept: config.accept ?? '',
+    }),
   }),
 ]);