+++ src/constants.ts
@@ -274,6 +274,7 @@ export const TEST_IDS = {
     fieldDatetimeMax: createSelector('data-testid column-editor field-datetime-max'),
     fieldCustomValues: createSelector('data-testid column-editor field-custom-values'),
     fieldLocalTime: createSelector('data-testid column-editor field-local-time'),
+    fieldFileAcceptTypes: createSelector('data-testid column-editor field-file-accept-types'),
   },
   dateEditor: {
     field: createSelector('data-testid date-editor field'),
@@ -291,6 +292,7 @@ export const TEST_IDS = {
     fieldSelect: createSelector('data-testid editableCell field-select'),
     fieldTextarea: createSelector('data-testid editableCell field-textarea'),
     fieldBoolean: createSelector('data-testid editableCell field-boolean'),
+    fieldFile: createSelector('data-testid editableCell field-file'),
   },
   tableActionsCell: {
     buttonStartEdit: createSelector('data-testid table-actions-cell button-start-edit'),